+++ DauCungDuocHotel/src/Entity/NhanVien.java
@@ -9,5 +9,68 @@ package Entity;
  * @author ps22004_taquocphong
  */
 public class NhanVien {
+    String maNV,tenNV,tenDN, Anh,matKhau,chucDanh;
+
+    public NhanVien() {
+    }
+
+    public NhanVien(String maNV, String tenNV, String tenDN, String Anh, String matKhau, String chucDanh) {
+	this.maNV = maNV;
+	this.tenNV = tenNV;
+	this.tenDN = tenDN;
+	this.Anh = Anh;
+	this.matKhau = matKhau;
+	this.chucDanh = chucDanh;
+    }
+
+    public String getMaNV() {
+	return maNV;
+    }
+
+    public void setMaNV(String maNV) {
+	this.maNV = maNV;
+    }
+
+    public String getTenNV() {
+	return tenNV;
+    }
+
+    public void setTenNV(String tenNV) {
+	this.tenNV = tenNV;
+    }
+
+    public String getTenDN() {
+	return tenDN;
+    }
+
+    public void setTenDN(String tenDN) {
+	this.tenDN = tenDN;
+    }
+
+    public String getAnh() {
+	return Anh;
+    }
+
+    public void setAnh(String Anh) {
+	this.Anh = Anh;
+    }
+
+    public String getMatKhau() {
+	return matKhau;
+    }
+
+    public void setMatKhau(String matKhau) {
+	this.matKhau = matKhau;
+    }
+
+    public String getChucDanh() {
+	return chucDanh;
+    }
+
+    public void setChucDanh(String chucDanh) {
+	this.chucDanh = chucDanh;
+    }
     
+    
+	    
 }
